refactor: batching module
