+++ orchestrator/src/worker/event_handler/triggers/batching/aggregator.rs
@@ -5,9 +5,9 @@ use crate::error::job::JobError;
 use crate::error::other::OtherError;
 use crate::types::batch::{AggregatorBatch, AggregatorBatchStatus, AggregatorBatchUpdates, AggregatorBatchWeights};
 use crate::utils::metrics::ORCHESTRATOR_METRICS;
-use crate::worker::event_handler::triggers::batching_v2::aggregator::AggregatorState::{Empty, NonEmpty};
-use crate::worker::event_handler::triggers::batching_v2::utils::{get_block_builtin_weights, get_block_version};
-use crate::worker::event_handler::triggers::batching_v2::BlockProcessingResult;
+use crate::worker::event_handler::triggers::batching::aggregator::AggregatorState::{Empty, NonEmpty};
+use crate::worker::event_handler::triggers::batching::utils::{get_block_builtin_weights, get_block_version};
+use crate::worker::event_handler::triggers::batching::BlockProcessingResult;
 use crate::worker::utils::biguint_vec_to_u8_vec;
 use bytes::Bytes;
 use chrono::{SubsecRound, Utc};
