+++ orchestrator/src/worker/event_handler/triggers/batching/snos.rs
@@ -2,8 +2,8 @@ use crate::core::config::{Config, ConfigParam, StarknetVersion};
 use crate::error::job::JobError;
 use crate::error::other::OtherError;
 use crate::types::batch::{SnosBatch, SnosBatchStatus, SnosBatchUpdates};
-use crate::worker::event_handler::triggers::batching_v2::utils::{get_block_builtin_weights, get_block_version};
-use crate::worker::event_handler::triggers::batching_v2::BlockProcessingResult;
+use crate::worker::event_handler::triggers::batching::utils::{get_block_builtin_weights, get_block_version};
+use crate::worker::event_handler::triggers::batching::BlockProcessingResult;
 use blockifier::bouncer::BouncerWeights;
 use chrono::{SubsecRound, Utc};
 use color_eyre::eyre::eyre;
